Changed from bot.py to main.py
