+++ main.py
@@ -10478,3 +10478,66 @@ class Section(discord.ui.Section):
             super().__init__(*new_children, **kwargs)
         else:
             super().__init__(*children, **kwargs)
+import random
+import discord
+from discord import app_commands
+from discord.ext import commands
+
+# Dictionary mapping specific emojis to cat reactions
+CAT_REACTIONS = {
+    # 🔴 Hated / Scary Items
+    "🥒": {"emoji": "🙀", "text": "The cat hisses, jumps high in the air and scrambles away."},
+    "🍋": {"emoji": "🙀", "text": "The cat takes a lick of the lemon and makes a shocked face."},
+    "💧": {"emoji": "😾", "text": "The cat gets wet and gives you a angry look."},
+    "🐶": {"emoji": "🤢", "text": "ew a dog."},
+    "🧹": {"emoji": "😾", "text": "The cat thinks it's a monster! The cat runs away!"},
+
+    # 🟢 Loved Items
+    "🐟": {"emoji": "😻", "text": "The cat purrs loudly, bites your arm and steals the fish!"},
+    "🥛": {"emoji": "😽", "text": "The cat drinks the milk and licks its whiskers in delight."},
+    "🧶": {"emoji": "😸", "text": "The cat jumps on the yarn and starts rolling around."},
+    "🍗": {"emoji": "😻", "text": "The cat meows excitedly and inhales down the chicken."},
+    "🌿": {"emoji": "🥴", "text": "Catnip! The cat rolls around, drools a little."},
+}
+
+# Random "Not Interested" responses for any other emoji/item
+UNINTERESTED_RESPONSES = [
+    "🐾 The cat gives it a brief sniff and completely ignores it.",
+    "🛋️ The cat looks at it and goes straight back to sleep.",
+    "😾 The cat scratches you.",
+    "🙄 The cat gives you a judgey look. It's not interested.",
+    "🐾 The cat walks away from you."
+]
+
+@app_commands.command(name="hisstest", description="Test how the cat reacts to a specific item or emoji!")
+@app_commands.describe(item="The emoji or item you want to test on the cat (e.g. 🥒, 🐟, 🧶)")
+async def hisstest(interaction: discord.Interaction, item: str):
+    # Clean up whitespace from input
+    clean_item = item.strip()
+    if not clean_item:
+        await interaction.response.send_message(
+            "Please provide an item or emoji."
+        )
+        return
+
+    # Check if the offered item exists in our special list
+    if clean_item in CAT_REACTIONS:
+        reaction = CAT_REACTIONS[clean_item]
+    clean_item = item.strip()
+    safe_item = discord.utils.escape_markdown(
+        discord.utils.escape_mentions(clean_item)
+    )
+
+    if clean_item in CAT_REACTIONS:
+        reaction = CAT_REACTIONS[clean_item]
+        response_text = f"**You offered:** {safe_item}\n**Cat's Reaction:** {reaction['emoji']} {reaction['text']}"
+    else:
+        # Pick a random "not interested" response for any other emoji
+        random_response = random.choice(UNINTERESTED_RESPONSES)
+        response_text = f"**You offered:** {safe_item}\n**Cat's Reaction:** {random_response}"
+
+    # Send the response back in Discord
+    await interaction.response.send_message(
+        response_text,
+        allowed_mentions=discord.AllowedMentions.none(),
+    )
